pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "product-scraper-template"
version = "0.1.0"
description = "Reusable template for web scraping projects."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Ryuhei", email = "you@example.com" }]
dependencies = [
    "requests",
    "beautifulsoup4",
    "PyYAML",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]
excel = [
    "pandas",
    "openpyxl",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["product_scraper*"]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = [
  "tests"
]

[tool.ruff]
line-length = 88
src = ["src", "tests"]

[project.scripts]
product-scraper = "product_scraper.cli:main"
